monday11

+++ src/main/java/io/springboot/Voda/Repository/MotorcoverRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import io.springboot.Voda.Entity.Motor_cover;
 
-public interface MotorcoverRepository extends JpaRepository<Motor_cover,Long>{
+public interface MotorcoverRepository extends JpaRepository<Motor_cover,Integer>{
 
 	Optional<Motor_cover> findById(Integer motorID);
